refactor task 5
